refactor: RestDocs 설정 부분 중복제거
